refactor(core)!: use fastbw to model selection
This commit uses the function rms::fastbw to model selection instead of MASS::stepAIC. The big difference is that the selections are more stable because we start with the full model and perform a backwards elimination.

Given this, the model selection happen successfully every time. The
execution time is longer but finite (hopefully).

The parameter `family` changes to be only a string ("gaussian" or "binomial") in order to differentiate the cases of rms::ols and rms::lrm.

Another consequence is that the parameters `steps` and `verbose` are unnecessary, and then removed.